requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
langchain
google-generativeai
python-dotenv
PyMuPDF
